refactor: #인공지능채팅 장원태 [240205] - 채팅방을 컴포넌트로 분리 - 채팅방 UI 및 디자인 작업중
