+++ src/main/reactjs/src/components/chat/chattingroom/ChatRoomHeader.js
@@ -1,0 +1,16 @@
+import React from 'react';
+
+const ChatRoomHeader = () => {
+    return (
+        <div>
+            <div className='fs_14 fw_500 col_blue2'>
+                <span>TODAC 채팅</span> {'>'} <span>상담받기</span>
+            </div>
+            <div className='fs_24 fw_700'>
+                TODAC 채팅
+            </div>
+        </div>
+    );
+};
+
+export default ChatRoomHeader;
